+++ prov/kubernetes/monitoring_job.go
@@ -53,8 +53,12 @@ func (o *actionOperator) ExecAction(ctx context.Context, cfg config.Configuratio
 	}
 	// Check if namespace was provided
 	var namespaceProvided bool
-	namespaceProvidedStr, ok := action.Data["providedNamespace"]
-	if b, err := strconv.ParseBool(namespaceProvidedStr); ok && err != nil {
+	namespaceProvidedStr, ok := action.Data["namespaceProvided"]
+	if ok {
+		b, err := strconv.ParseBool(namespaceProvidedStr)
+		if err != nil {
+			return true, errors.New(`unable to transform namespaceProvided value to boolean`)
+		}
 		namespaceProvided = b
 	}